@@ -2,15 +2,23 @@
 
 namespace App\Drivers\Source\Renins;
 
+use DateTime;
 use Illuminate\Contracts\Support\Arrayable;
 use Illuminate\Support\Carbon;
 
+/**
+ * Class ReninsCalcCollector
+ * @package App\Drivers\Source\Renins
+ */
 class ReninsCalcCollector implements Arrayable
 {
     protected const TIME_FORMAT = 'Y-m-d\TH:i:s.v\Z';
 
     protected array $data = [];
 
+    /**
+     * ReninsCalcCollector constructor.
+     */
     public function __construct()
     {
         $this->data = [
@@ -63,14 +71,20 @@ class ReninsCalcCollector implements Arrayable
         ];
     }
 
-    public function setStartEnd($dateStart, $dateEnd): void
+    /**
+     * @param $dateStart
+     * @param $dateEnd
+     */
+    public function setContractStartEnd(string $policyStartDate, string $policyEndDate): void
     {
-        $this->data['dateBeg'] = $this->toTime($dateStart);
-        $this->data['dateEnd'] = $this->toTime($dateEnd);
+        $this->data['dateBeg'] = $this->toTime($policyStartDate);
+        $this->data['dateEnd'] = $this->toTime($policyEndDate);
     }
 
-    // только для life
-    public function workStatus(array $subject)
+    /**
+     * @param array $subject
+     */
+    public function workStatus(array $subject): void
     {
         $this->data['parameters']['parameters'][] = [
             'name' => 'Статус занятости Страхователя',
@@ -80,8 +94,10 @@ class ReninsCalcCollector implements Arrayable
         ];
     }
 
-    // только для life
-    public function subjectIsObject(bool $is = true)
+    /**
+     * @param bool $is
+     */
+    public function subjectIsObject(bool $is = true): void
     {
         $this->data['parameters']['parameters'][] = [
             'name' => 'Застрахованный является Страхователем',
@@ -91,8 +107,10 @@ class ReninsCalcCollector implements Arrayable
         ];
     }
 
-    // только property
-    public function setBuildDate(int $year)
+    /**
+     * @param int $year
+     */
+    public function setBuildDate(int $year): void
     {
         $this->data['parameters']['parameters'][] = [
             'name' => 'Год постройки',
@@ -102,14 +120,18 @@ class ReninsCalcCollector implements Arrayable
         ];
     }
 
-    public static function getFormatPhone($phone) {
+    /**
+     * @param $phone
+     */
+    public static function getFormatPhone(string $phone): ?string
+    {
         preg_match(
             '/(\d{1,4})\s*\(?(\d{3,5})\)?\s*(\d{3})[\s-]?(\d{2})[\s-]?(\d{2})/',
             $phone,
             $matches
         );
         if (count($matches) !== 6) {
-            return false;
+            return null;
         }
 
         array_shift($matches);
@@ -118,7 +140,10 @@ class ReninsCalcCollector implements Arrayable
         return $phone;
     }
 
-    public function setBirthDate(string $date)
+    /**
+     * @param string $date
+     */
+    public function setBirthDate(string $date): void
     {
         $this->data['insurant']['physical'] = [
             'birthDate' => $this->toTime($date)
@@ -131,7 +156,9 @@ class ReninsCalcCollector implements Arrayable
         ];
     }
 
-    // только для life
+    /**
+     * @param int $sex
+     */
     public function setSex(int $sex): void
     {
         $this->data['parameters']['parameters'][] = [
@@ -142,6 +169,9 @@ class ReninsCalcCollector implements Arrayable
         ];
     }
 
+    /**
+     * @param string $city
+     */
     public function setCreditCity(string $city): void
     {
         $this->data['parameters']['parameters'][] = [
@@ -152,6 +182,9 @@ class ReninsCalcCollector implements Arrayable
         ];
     }
 
+    /**
+     * @param float $sum
+     */
     public function setCreditSum(float $sum): void
     {
         $this->data['parameters']['parameters'][] = [
@@ -162,6 +195,9 @@ class ReninsCalcCollector implements Arrayable
         ];
     }
 
+    /**
+     * @param string $bik
+     */
     public function setBankBik(string $bik): void
     {
         $this->data['parameters']['parameters'][] = [
@@ -172,6 +208,10 @@ class ReninsCalcCollector implements Arrayable
         ];
     }
 
+    /**
+     * @param array $risks
+     * @param string $alias
+     */
     public function addObject(array $risks, string $alias): void
     {
         $this->data['insuranceObjects']['objects'][] =[
@@ -184,6 +224,10 @@ class ReninsCalcCollector implements Arrayable
         ];
     }
 
+    /**
+     * @param $date
+     * @return string|null
+     */
     protected function toTime($date): ?string
     {
         if (is_numeric($date)) {
@@ -192,13 +236,16 @@ class ReninsCalcCollector implements Arrayable
         if (is_string($date)) {
             return Carbon::parse($date)->format(self::TIME_FORMAT);
         }
-        if ($date instanceof \DateTime) {
+        if ($date instanceof DateTime) {
             return $date->format(self::TIME_FORMAT);
         }
 
         return null;
     }
 
+    /**
+     * @return array
+     */
     public function toArray()
     {
         return [
