@@ -2,6 +2,12 @@
 
 namespace App\Drivers\Source\Renins;
 
+use Arr;
+
+/**
+ * Class ReninsCreateCollector
+ * @package App\Drivers\Source\Renins
+ */
 class ReninsCreateCollector extends ReninsCalcCollector
 {
     const CITIZENSHIP = 'Россия';
@@ -11,6 +17,9 @@ class ReninsCreateCollector extends ReninsCalcCollector
     const CREDIT_CITY = 'Москва';
     const CURRENCY = 'RUR';
 
+    /**
+     * ReninsCreateCollector constructor.
+     */
     public function __construct()
     {
         parent::__construct();
@@ -97,7 +106,11 @@ class ReninsCreateCollector extends ReninsCalcCollector
         ];
     }
 
-    public function setPayPlan($dateStart, float $sum)
+    /**
+     * @param $dateStart
+     * @param float $sum
+     */
+    public function setPayPlan(string $dateStart, float $sum): void
     {
         $this->data['paymentsPlan']['payments'][] = [
             'number' => 1,
@@ -107,8 +120,18 @@ class ReninsCreateCollector extends ReninsCalcCollector
         ];
     }
 
-    public function setPropertyAddress($state, $city, $street, $house, $kladr)
-    {
+    /**
+     * @param string $state
+     * @param string $city
+     * @param string $street
+     * @param string $house
+     */
+    public function setPropertyAddress(
+        string $state,
+        string $city,
+        string $street,
+        string $house
+    ): void {
         $this->data['parameters']['parameters'][] = [
             'name' => 'Строковое представление адреса',
             'code' => 'dogovor.adresImushestva.adresStroka',
@@ -126,7 +149,10 @@ class ReninsCreateCollector extends ReninsCalcCollector
         ];
     }
 
-    public function setKladr(string $kladr)
+    /**
+     * @param string $kladr
+     */
+    public function setKladr(string $kladr): void
     {
         $this->data['parameters']['parameters'][] = [
             'name' => 'Код КЛАДР',
@@ -136,29 +162,39 @@ class ReninsCreateCollector extends ReninsCalcCollector
         ];
     }
 
-    public function setStartEnd($dateStart, $dateEnd): void
+    /**
+     * @param string $dateStart
+     * @param string $dateEnd
+     */
+    public function setContractStartEnd(string $dateStart, string $dateEnd): void
     {
         $this->data['dateBeg'] = $this->toTime($dateStart);
         $this->data['dateEnd'] = $this->toTime($dateEnd);
     }
 
+    /**
+     * @param float $sum
+     */
     public function setCreditSum(float $sum): void
     {
         $this->data['parameters']['parameters'][] = [
             'name' => 'Страховая стоимость объекта',
             'code' => 'dogovor.strahStoimostObjekta',
             'type' => 'Вещественный',
-            'decimalValue' => 3000000,
+            'decimalValue' => $sum,
         ];
         $this->data['parameters']['parameters'][] = [
             'name' => 'Сумма кредита',
             'code' => 'dogovor.ipotechnDogovor.summaKredita',
             'type' => 'Вещественный',
-            'decimalValue' => 3000000,
+            'decimalValue' => $sum,
         ];
     }
 
-    public function setCreditNumber(string $number)
+    /**
+     * @param string $number
+     */
+    public function setCreditNumber(string $number): void
     {
         $this->data['parameters']['parameters'][] = [
             'name' => 'Номер',
@@ -168,7 +204,10 @@ class ReninsCreateCollector extends ReninsCalcCollector
         ];
     }
 
-    public function setBirthDate(string $date)
+    /**
+     * @param string $date
+     */
+    public function setBirthDate(string $date): void
     {
         $this->data['parameters']['parameters'][] = [
             'name' => 'Дата рождения',
@@ -178,7 +217,10 @@ class ReninsCreateCollector extends ReninsCalcCollector
         ];
     }
 
-    public function setBirthDateSubject(string $date)
+    /**
+     * @param string $date
+     */
+    public function setBirthDateSubject(string $date): void
     {
         $this->data['insurant']['physical'] = array_merge(
             $this->data['insurant']['physical'],
@@ -186,23 +228,26 @@ class ReninsCreateCollector extends ReninsCalcCollector
         );
     }
 
-    public function setHumanInfo(array $subject)
+    /**
+     * @param array $subject
+     */
+    public function setHumanInfo(array $subject): void
     {
         $this->data['insurant']['physical'] = [
-            'lastName' => \Arr::get($subject, 'lastName'),
-            'firstName' => \Arr::get($subject, 'firstName'),
-            'middleName' => \Arr::get($subject, 'middleName'),
-            'birthDate' => $this->toTime(\Arr::get($subject, 'birthDate')),
-            'email' => \Arr::get($subject, 'email'),
-            'phone' => self::getFormatPhone(\Arr::get($subject, 'phone')),
-            'sex' => \Arr::get($subject, 'gender', 0) ? "F" : "M",
+            'lastName' => Arr::get($subject, 'lastName'),
+            'firstName' => Arr::get($subject, 'firstName'),
+            'middleName' => Arr::get($subject, 'middleName'),
+            'birthDate' => $this->toTime(Arr::get($subject, 'birthDate')),
+            'email' => Arr::get($subject, 'email'),
+            'phone' => self::getFormatPhone(Arr::get($subject, 'phone')),
+            'sex' => Arr::get($subject, 'gender', 0) ? "F" : "M",
             'citizenship' => self::CITIZENSHIP,
             'document' => [
                 'type' => self::DOC_TYPE,
-                'series' => \Arr::get($subject, 'docSeries'),
-                'number' => \Arr::get($subject, 'docNumber'),
-                'placeOfIssue' => \Arr::get($subject, 'docIssuePlace'),
-                'dateOfIssue' => $this->toTime(\Arr::get($subject, 'docIssueDate')),
+                'series' => Arr::get($subject, 'docSeries'),
+                'number' => Arr::get($subject, 'docNumber'),
+                'placeOfIssue' => Arr::get($subject, 'docIssuePlace'),
+                'dateOfIssue' => $this->toTime(Arr::get($subject, 'docIssueDate')),
                 'kodPodrazd' => '',
             ],
             'factAddress' => [
@@ -224,23 +269,26 @@ class ReninsCreateCollector extends ReninsCalcCollector
                         ' к.',
                     ],
                     '',
-                    \Arr::get($subject, 'state')
+                    Arr::get($subject, 'state')
                 ), ".\n\r, \t\0"),
                 'addressText' => implode(
                     ', ',
                     array_filter([
                         self::POST_INDEX_MOSCOW,
                         self::COUNTRY,
-                        \Arr::get($subject, 'city'),
-                        'ул. '  . \Arr::get($subject, 'street'),
-                        \Arr::get($subject, 'house'),
+                        Arr::get($subject, 'city'),
+                        'ул. '  . Arr::get($subject, 'street'),
+                        Arr::get($subject, 'house'),
                     ])
                 ),
             ],
         ];
     }
 
-    public function toArray()
+    /**
+     * @return array
+     */
+    public function toArray(): array
     {
         return [
             'policy' => $this->data
