@@ -6,14 +6,21 @@ use Ajaxray\PHPWatermark\Watermark;
 use App\Models\Contracts;
 use App\Services\WatermarkService;
 use Barryvdh\DomPDF\Facade as PDF;
-use http\Exception\RuntimeException;
+use RuntimeException;
 use setasign\Fpdi\FpdfTpl;
 use setasign\Fpdi\Fpdi;
 
+/**
+ * Trait PrintPdfTrait
+ * @package App\Drivers\Traits
+ */
 trait PrintPdfTrait
 {
-    use LoggerTrait;
 
+    /**
+     * @param $path
+     * @return string
+     */
     public static function generateBase64($path): string
     {
         return base64_encode(file_get_contents($path));
@@ -49,7 +56,7 @@ trait PrintPdfTrait
         $dir = substr($filename, 0, strrpos($filename, DIRECTORY_SEPARATOR) + 1);
         if (!file_exists($dir)) {
             if (!mkdir($dir, 0777, true) && !is_dir($dir)) {
-                throw new \RuntimeException(sprintf('Directory "%s" was not created :' . $dir, $dir));
+                throw new RuntimeException('Directory "%s" was not created :' . $dir);
             }
         }
         $pdf->save($filename);
@@ -57,17 +64,29 @@ trait PrintPdfTrait
         return $filename;
     }
 
+    /**
+     * @param string $path
+     * @param string $content
+     * @return bool
+     */
     protected static function saveFileFromContent(string $path, string $content): bool
     {
         $dir = substr($path, 0, strrpos($path, DIRECTORY_SEPARATOR));
         if (!file_exists($dir)) {
             if (!mkdir($dir, 0777, true) && !is_dir($dir)) {
-                throw new \RuntimeException(sprintf('Directory "%s" was not created', $dir));
+                throw new RuntimeException("Directory '$dir' was not created");
             }
         }
         return (bool)file_put_contents($path, $content);
     }
 
+    /**
+     * @param $pathPdf
+     * @param $pathWatermark
+     * @param $resultPath
+     * @return string
+     * @throws \Exception
+     */
     public static function setWatermark($pathPdf, $pathWatermark, $resultPath): string
     {
         $watermark = new WatermarkService($pathPdf, true);
@@ -75,7 +94,7 @@ trait PrintPdfTrait
         $pdf = $watermark->withImage($pathWatermark, $resultPath);
 
         if (!$pdf) {
-            self::abortLog('Not install imagic or not set rules for pdf', RuntimeException::class);
+            throw new RuntimeException('Not install imagic or not set rules for pdf');
         }
 
         return $pdf;
