@@ -2,7 +2,9 @@
 
 namespace App\Exceptions\Drivers;
 
+use Symfony\Component\HttpFoundation\Response;
+
 class AlphaException extends \Exception implements DriverExceptionInterface
 {
-
+    protected $code = Response::HTTP_NOT_ACCEPTABLE;
 }
