@@ -10,11 +10,11 @@ use OpenApi\Annotations as OA;
  *     required={"activeFrom", "activeTo", "programCode", "remainingDebt", "objects", "isOwnership"},
  *     schema="CalculateRequest",
  *     @OA\Property(property="programCode", type="string", example="RENSINS_MORTGAGE_002_01", description="Код программы"),
- *     @OA\Property(property="activeFrom", type="date", example="2021-05-21", description="Дата начала действия договора Ипотеки"),
- *     @OA\Property(property="activeTo", type="date", example="2022-05-20", description="Дата окончания действия договора Ипотеки"),
+ *     @OA\Property(property="activeFrom", type="date", example="2021-06-21", description="Дата начала действия договора Ипотеки"),
+ *     @OA\Property(property="activeTo", type="date", example="2022-06-20", description="Дата окончания действия договора Ипотеки"),
  *     @OA\Property(property="remainingDebt", type="float", example=1500000, description="Остаток  задолженности по договору ипотеки"),
  *     @OA\Property(property="isOwnership", type="boolean", example=1, description="Признак наличия права собственности"),
- *     @OA\Property(property="mortgageeBank", type="string", example="ПАО Сбербанк", description="Банк-залогодержатель (выгодоприобретатель)"),
+ *     @OA\Property(property="mortgageeBank", type="string", example="СБЕРБАНК РОССИИ", description="Банк-залогодержатель (выгодоприобретатель)"),
  *     @OA\Property(property="objects", description="Объекты страхования", type="object",
  *        @OA\Property(property="property", type="object", description="Тип объекта страхования - Недвижимое имущество",
  *            required={"buildYear"},
