@@ -10,12 +10,12 @@ use Illuminate\Validation\Rule;
  *     schema="CreatePolicyRequest",
  *     @OA\Property(property="programCode", type="string", example="RENSINS_MORTGAGE_002_01", description="Код программы страхования"),
  *     @OA\Property(property="ownerCode", type="string", example="STRAHOVKA", description="Код канала, откуда идут запросы для методов"),
- *     @OA\Property(property="activeFrom", type="date", example="2021-05-18", description="Дата начала действия договора страхования"),
- *     @OA\Property(property="activeTo", type="date", example="2022-05-17", description="Дата окончания действия договора страхования"),
+ *     @OA\Property(property="activeFrom", type="date", example="2021-06-18", description="Дата начала действия договора страхования"),
+ *     @OA\Property(property="activeTo", type="date", example="2022-06-17", description="Дата окончания действия договора страхования"),
  *     @OA\Property(property="remainingDebt", type="integer", example=100000, description="Страховая сумма"),
  *     @OA\Property(property="mortgageAgreementNumber", type="string", example="100000", description="Номер ипотечного договора"),
  *     @OA\Property(property="isOwnership", type="boolean", example=1, description="Указатель на наличие в собственности"),
- *     @OA\Property(property="mortgageeBank", type="string", example="ПАО Сбербанк", description="Банк выгодоприобретатель"),
+ *     @OA\Property(property="mortgageeBank", type="string", example="СБЕРБАНК РОССИИ", description="Банк выгодоприобретатель"),
  *     @OA\Property(property="objects", ref="#/components/schemas/Objects", description="Объекты страхования"),
  *     @OA\Property(property="subject", ref="#/components/schemas/Subject", description="Субъект страхования"),
  * )
