@@ -2,9 +2,9 @@
 
 /**
  * @OA\Info(
- *     title="Телемед адаптер",
+ *     title="Ипотека адаптер",
  *     version="0.1",
- *     description="Адаптер расчета и оформления Телемед-полисов на стороне Страховка.Ру",
+ *     description="Адаптер расчета и оформления Ипотека-полисов на стороне Страховка.Ру",
  *     @OA\Contact(
  *         email="info@strahovka.ru"
  *     )
