@@ -7,6 +7,10 @@ use Artisaninweb\SoapWrapper\SoapWrapper;
 use Illuminate\Support\ServiceProvider;
 use Strahovka\Payment\PayService;
 
+/**
+ * Class DriverServiceProvider
+ * @package App\Providers
+ */
 class DriverServiceProvider extends ServiceProvider
 {
     /**
@@ -19,9 +23,15 @@ class DriverServiceProvider extends ServiceProvider
         $this->app->singleton(DriverService::class, function ($app) {
             return new DriverService();
         });
+        $drivers = config('mortgage.drivers');
+        foreach ($drivers as $code => $driver) {
+            $this->app->singleton($code, function($app) use ($driver, $code) {
+                return new $driver(config(), "mortgage.{$code}.");
+            });
+        }
 
         $this->app->bind(PayService::class, function ($app) {
-            if (in_array(env('APP_ENV'), ['local', 'testing'])) {
+            if (in_array(config('app.env'), ['local', 'testing'])) {
                 return new class (new SoapWrapper()) extends PayService {
                     public function __construct(SoapWrapper $soapWrapper) {
                         parent::__construct($soapWrapper);
