@@ -1,243 +1,247 @@
 <?php
 
 return [
-    'api' => [
-        /*
-        |--------------------------------------------------------------------------
-        | Edit to set the api's title
-        |--------------------------------------------------------------------------
-        */
+    'default' => 'default',
+    'documentations' => [
+        'default' => [
+            'api' => [
+                'title' => 'Адаптер ипотека',
+            ],
 
-        'title' => 'Адаптер телемедицина',
-    ],
+            'routes' => [
+                /*
+                 * Route for accessing api documentation interface
+                */
+                'api' => 'api/docs',
+            ],
+            'paths' => [
+                /*
+                 * File name of the generated json documentation file
+                */
+                'docs_json' => 'api-docs.json',
 
-    'routes' => [
-        /*
-        |--------------------------------------------------------------------------
-        | Route for accessing api documentation interface
-        |--------------------------------------------------------------------------
-        */
+                /*
+                 * File name of the generated YAML documentation file
+                */
+                'docs_yaml' => 'api-docs.yaml',
 
-        'api' => 'api/docs',
+                /*
+                * Set this to `json` or `yaml` to determine which documentation file to use in UI
+                */
+                'format_to_use_for_docs' => env('L5_FORMAT_TO_USE_FOR_DOCS', 'json'),
 
-        /*
-        |--------------------------------------------------------------------------
-        | Route for accessing parsed swagger annotations.
-        |--------------------------------------------------------------------------
-        */
+                /*
+                 * Absolute paths to directory containing the swagger annotations are stored.
+                */
+                'annotations' => [
+                    base_path('app'),
+                ],
 
-        'docs' => 'docs',
-
-        /*
-        |--------------------------------------------------------------------------
-        | Route for Oauth2 authentication callback.
-        |--------------------------------------------------------------------------
-        */
-
-        'oauth2_callback' => 'api/oauth2-callback',
-
-        /*
-        |--------------------------------------------------------------------------
-        | Middleware allows to prevent unexpected access to API documentation
-        |--------------------------------------------------------------------------
-         */
-        'middleware' => [
-            'api' => [],
-            'asset' => [],
-            'docs' => [],
-            'oauth2_callback' => [],
+            ],
         ],
     ],
+    'defaults' => [
+        'routes' => [
+            /*
+             * Route for accessing parsed swagger annotations.
+            */
+            'docs' => 'docs',
 
-    'paths' => [
-        /*
-        |--------------------------------------------------------------------------
-        | Absolute path to location where parsed swagger annotations will be stored
-        |--------------------------------------------------------------------------
-        */
+            /*
+             * Route for Oauth2 authentication callback.
+            */
+            'oauth2_callback' => 'api/oauth2-callback',
 
-        'docs' => storage_path('api-docs'),
+            /*
+             * Middleware allows to prevent unexpected access to API documentation
+            */
+            'middleware' => [
+                'api' => [],
+                'asset' => [],
+                'docs' => [],
+                'oauth2_callback' => [],
+            ],
 
-        /*
-        |--------------------------------------------------------------------------
-        | File name of the generated json documentation file
-        |--------------------------------------------------------------------------
-        */
+            /*
+             * Route Group options
+            */
+            'group_options' => [],
+        ],
 
-        'docs_json' => 'api-docs.json',
+        'paths' => [
+            /*
+             * Absolute path to location where parsed annotations will be stored
+            */
+            'docs' => storage_path('api-docs'),
 
-        /*
-        |--------------------------------------------------------------------------
-        | File name of the generated YAML documentation file
-        |--------------------------------------------------------------------------
-         */
+            /*
+             * Absolute path to directory where to export views
+            */
+            'views' => base_path('resources/views/vendor/l5-swagger'),
 
-        'docs_yaml' => 'api-docs.yaml',
+            /*
+             * Edit to set the api's base path
+            */
+            'base' => env('L5_SWAGGER_BASE_PATH', null),
 
-        /*
-        |--------------------------------------------------------------------------
-        | Absolute paths to directory containing the swagger annotations are stored.
-        |--------------------------------------------------------------------------
-        */
+            /*
+             * Edit to set path where swagger ui assets should be stored
+            */
+            'swagger_ui_assets_path' => env('L5_SWAGGER_UI_ASSETS_PATH', 'vendor/swagger-api/swagger-ui/dist/'),
 
-        'annotations' => [
-            base_path('app/Http'),
+            /*
+             * Absolute path to directories that should be exclude from scanning
+             * @deprecated Please use `scanOptions.exclude`
+             * `scanOptions.exclude` overwrites this
+            */
+            'excludes' => [],
+        ],
+
+        'scanOptions' => [
+            /**
+             * analyser: defaults to \OpenApi\StaticAnalyser .
+             * @see \OpenApi\scan
+             */
+            'analyser' => null,
+
+            /**
+             * analysis: defaults to a new \OpenApi\Analysis .
+             * @see \OpenApi\scan
+             */
+            'analysis' => null,
+
+            /**
+             * Custom query path processors classes.
+             *
+             * @link https://github.com/zircote/swagger-php/tree/master/Examples/schema-query-parameter-processor
+             * @see \OpenApi\scan
+             */
+            'processors' => [
+                // new \App\SwaggerProcessors\SchemaQueryParameter(),
+            ],
+
+            /**
+             * pattern: string       $pattern File pattern(s) to scan (default: *.php) .
+             * @see \OpenApi\scan
+             */
+            'pattern' => null,
+
+            /*
+             * Absolute path to directories that should be exclude from scanning
+             * @note This option overwrites `paths.excludes`
+             * @see \OpenApi\scan
+            */
+            'exclude' => [],
         ],
 
         /*
-        |--------------------------------------------------------------------------
-        | Absolute path to directory where to export views
-        |--------------------------------------------------------------------------
+         * API security definitions. Will be generated into documentation file.
         */
+        'securityDefinitions' => [
+            'securitySchemes' => [
+                /*
+                 * Examples of Security schemes
+                */
+                /*
+                'api_key_security_example' => [ // Unique name of security
+                    'type' => 'apiKey', // The type of the security scheme. Valid values are "basic", "apiKey" or "oauth2".
+                    'description' => 'A short description for security scheme',
+                    'name' => 'api_key', // The name of the header or query parameter to be used.
+                    'in' => 'header', // The location of the API key. Valid values are "query" or "header".
+                ],
+                'oauth2_security_example' => [ // Unique name of security
+                    'type' => 'oauth2', // The type of the security scheme. Valid values are "basic", "apiKey" or "oauth2".
+                    'description' => 'A short description for oauth2 security scheme.',
+                    'flow' => 'implicit', // The flow used by the OAuth2 security scheme. Valid values are "implicit", "password", "application" or "accessCode".
+                    'authorizationUrl' => 'http://example.com/auth', // The authorization URL to be used for (implicit/accessCode)
+                    //'tokenUrl' => 'http://example.com/auth' // The authorization URL to be used for (password/application/accessCode)
+                    'scopes' => [
+                        'read:projects' => 'read your projects',
+                        'write:projects' => 'modify projects in your account',
+                    ]
+                ],
+                */
 
-        'views' => base_path('resources/views/vendor/l5-swagger'),
+                /* Open API 3.0 support
+                'passport' => [ // Unique name of security
+                    'type' => 'oauth2', // The type of the security scheme. Valid values are "basic", "apiKey" or "oauth2".
+                    'description' => 'Laravel passport oauth2 security.',
+                    'in' => 'header',
+                    'scheme' => 'https',
+                    'flows' => [
+                        "password" => [
+                            "authorizationUrl" => config('app.url') . '/oauth/authorize',
+                            "tokenUrl" => config('app.url') . '/oauth/token',
+                            "refreshUrl" => config('app.url') . '/token/refresh',
+                            "scopes" => []
+                        ],
+                    ],
+                ],
+                */
+            ],
+            'security' => [
+                /*
+                 * Examples of Securities
+                */
+                [
+                    /*
+                    'oauth2_security_example' => [
+                        'read',
+                        'write'
+                    ],
 
-        /*
-        |--------------------------------------------------------------------------
-        | Edit to set the api's base path
-        |--------------------------------------------------------------------------
-        */
-
-        'base' => env('L5_SWAGGER_BASE_PATH', null),
-
-        /*
-        |--------------------------------------------------------------------------
-        | Edit to set path where swagger ui assets should be stored
-        |--------------------------------------------------------------------------
-        */
-
-        'swagger_ui_assets_path' => env('L5_SWAGGER_UI_ASSETS_PATH', 'vendor/swagger-api/swagger-ui/dist/'),
-
-        /*
-        |--------------------------------------------------------------------------
-        | Absolute path to directories that you would like to exclude from swagger generation
-        |--------------------------------------------------------------------------
-        */
-
-        'excludes' => [
-            storage_path(),
-            base_path()."/tests",
-            base_path()."/resources/views",
-            base_path()."/config",
-            base_path()."/vendor"
-        ],
-    ],
-
-    /*
-    |--------------------------------------------------------------------------
-    | API security definitions. Will be generated into documentation file.
-    |--------------------------------------------------------------------------
-    */
-    'security' => [
-        /*
-        |--------------------------------------------------------------------------
-        | Examples of Security definitions
-        |--------------------------------------------------------------------------
-        */
-        /*
-        'api_key_security_example' => [ // Unique name of security
-            'type' => 'apiKey', // The type of the security scheme. Valid values are "basic", "apiKey" or "oauth2".
-            'description' => 'A short description for security scheme',
-            'name' => 'api_key', // The name of the header or query parameter to be used.
-            'in' => 'header', // The location of the API key. Valid values are "query" or "header".
-        ],
-        'oauth2_security_example' => [ // Unique name of security
-            'type' => 'oauth2', // The type of the security scheme. Valid values are "basic", "apiKey" or "oauth2".
-            'description' => 'A short description for oauth2 security scheme.',
-            'flow' => 'implicit', // The flow used by the OAuth2 security scheme. Valid values are "implicit", "password", "application" or "accessCode".
-            'authorizationUrl' => 'http://example.com/auth', // The authorization URL to be used for (implicit/accessCode)
-            //'tokenUrl' => 'http://example.com/auth' // The authorization URL to be used for (password/application/accessCode)
-            'scopes' => [
-                'read:projects' => 'read your projects',
-                'write:projects' => 'modify projects in your account',
-            ]
-        ],
-        */
-
-        /* Open API 3.0 support
-        'passport' => [ // Unique name of security
-            'type' => 'oauth2', // The type of the security scheme. Valid values are "basic", "apiKey" or "oauth2".
-            'description' => 'Laravel passport oauth2 security.',
-            'in' => 'header',
-            'scheme' => 'https',
-            'flows' => [
-                "password" => [
-                    "authorizationUrl" => config('app.url') . '/oauth/authorize',
-                    "tokenUrl" => config('app.url') . '/oauth/token',
-                    "refreshUrl" => config('app.url') . '/token/refresh',
-                    "scopes" => []
+                    'passport' => []
+                    */
                 ],
             ],
         ],
+
+        /*
+         * Set this to `true` in development mode so that docs would be regenerated on each request
+         * Set this to `false` to disable swagger generation on production
         */
-    ],
+        'generate_always' => env('L5_SWAGGER_GENERATE_ALWAYS', false),
 
-    /*
-    |--------------------------------------------------------------------------
-    | Turn this off to remove swagger generation on production
-    |--------------------------------------------------------------------------
-    */
+        /*
+         * Set this to `true` to generate a copy of documentation in yaml format
+        */
+        'generate_yaml_copy' => env('L5_SWAGGER_GENERATE_YAML_COPY', false),
 
-    'generate_always' => env('L5_SWAGGER_GENERATE_ALWAYS', false),
+        /*
+         * Edit to trust the proxy's ip address - needed for AWS Load Balancer
+         * string[]
+        */
+        'proxy' => false,
 
-    /*
-    |--------------------------------------------------------------------------
-    | Turn this on to generate a copy of documentation in yaml format
-    |--------------------------------------------------------------------------
-     */
+        /*
+         * Configs plugin allows to fetch external configs instead of passing them to SwaggerUIBundle.
+         * See more at: https://github.com/swagger-api/swagger-ui#configs-plugin
+        */
+        'additional_config_url' => null,
 
-    'generate_yaml_copy' => env('L5_SWAGGER_GENERATE_YAML_COPY', false),
+        /*
+         * Apply a sort to the operation list of each API. It can be 'alpha' (sort by paths alphanumerically),
+         * 'method' (sort by HTTP method).
+         * Default is the order returned by the server unchanged.
+        */
+        'operations_sort' => env('L5_SWAGGER_OPERATIONS_SORT', null),
 
-    /*
-    |--------------------------------------------------------------------------
-    | Edit to set the swagger version number
-    |--------------------------------------------------------------------------
-    */
+        /*
+         * Pass the validatorUrl parameter to SwaggerUi init on the JS side.
+         * A null value here disables validation.
+        */
+        'validator_url' => null,
 
-    'swagger_version' => env('SWAGGER_VERSION', '3.0'),
+        /*
+         * Persist authorization login after refresh browser
+         */
+        'persist_authorization' => true,
 
-    /*
-    |--------------------------------------------------------------------------
-    | Edit to trust the proxy's ip address - needed for AWS Load Balancer
-    |--------------------------------------------------------------------------
-    */
-
-    'proxy' => false,
-
-    /*
-    |--------------------------------------------------------------------------
-    | Configs plugin allows to fetch external configs instead of passing them to SwaggerUIBundle.
-    | See more at: https://github.com/swagger-api/swagger-ui#configs-plugin
-    |--------------------------------------------------------------------------
-    */
-
-    'additional_config_url' => null,
-
-    /*
-    |--------------------------------------------------------------------------
-    | Apply a sort to the operation list of each API. It can be 'alpha' (sort by paths alphanumerically),
-    | 'method' (sort by HTTP method).
-    | Default is the order returned by the server unchanged.
-    |--------------------------------------------------------------------------
-    */
-
-    'operations_sort' => env('L5_SWAGGER_OPERATIONS_SORT', null),
-
-    /*
-    |--------------------------------------------------------------------------
-    | Uncomment to pass the validatorUrl parameter to SwaggerUi init on the JS
-    | side.  A null value here disables validation.
-    |--------------------------------------------------------------------------
-    */
-
-    'validator_url' => null,
-
-    /*
-    |--------------------------------------------------------------------------
-    | Uncomment to add constants which can be used in annotations
-    |--------------------------------------------------------------------------
-     */
-    'constants' => [
-        'L5_SWAGGER_CONST_HOST' => env('L5_SWAGGER_CONST_HOST', 'http://my-default-host.com'),
+        /*
+         * Uncomment to add constants which can be used in annotations
+         */
+        // 'constants' => [
+        // 'L5_SWAGGER_CONST_HOST' => env('L5_SWAGGER_CONST_HOST', 'http://my-default-host.com'),
+        // ],
     ],
 ];
